move error classes to their own file
