+++ rbc/rbclib/debug_allocator.py
@@ -1,23 +1,5 @@
 from ._rbclib import lib, ffi
-
-
-class DebugAllocatorError(Exception):
-    pass
-
-
-class InvalidFreeError(DebugAllocatorError):
-    pass
-
-
-class MemoryLeakError(DebugAllocatorError):
-
-    def __init__(self, leaks):
-        lines = [f'Found {len(leaks)} memory leaks:']
-        for addr, seq in leaks:
-            lines.append(f'    {addr} (seq = {seq})')
-        message = '\n'.join(lines)
-        super().__init__(message)
-        self.leaks = leaks
+from .errors import DebugAllocatorError, InvalidFreeError, MemoryLeakError
 
 
 class DebugAllocator:
